+++ Tests/Runtime/ConfigManagerTests.cs
@@ -9,18 +9,85 @@ using UnityEngine;
 using System.Linq;
 using System.IO;
 using System.Net.NetworkInformation;
+using System.Threading.Tasks;
+using Unity.Services.Authentication;
+using Unity.Services.Core;
+using UnityEditor;
 
 namespace Unity.RemoteConfig.Tests
 {
     internal class ConfigManagerTests
     {
+        private SerializedObject _projectSettingsObject;
+        private SerializedProperty _cloudProjectIdProperty;
+        private SerializedProperty _cloudProjectNameProperty;
+        private SerializedProperty _versionProperty;
+        private string _previousProjectId;
+        private string _previousProjectName;
+        private string _previousVersion;
+        
+        private void FixYamatoProjectSettings()
+        {
+            // Cloud Project ID needs to be linked or the SDK will fail to start.
+            // Since this cannot be set in Yamato's transient test projects, we need to do a little hackery...
+            const string ProjectSettingsAssetPath = "ProjectSettings/ProjectSettings.asset";
+            _projectSettingsObject = new SerializedObject(UnityEditor.AssetDatabase.LoadAllAssetsAtPath(ProjectSettingsAssetPath)[0]);  
+            _cloudProjectIdProperty = _projectSettingsObject.FindProperty("cloudProjectId");
+            _cloudProjectNameProperty = _projectSettingsObject.FindProperty("projectName");
+            _versionProperty = _projectSettingsObject.FindProperty("bundleVersion"); // NOTE: this is Project Settings -> Player -> Version
+            
+            _previousProjectId = _cloudProjectIdProperty.stringValue;
+            _previousProjectName = _cloudProjectNameProperty.stringValue;
+            _previousVersion = _versionProperty.stringValue;
+            _cloudProjectIdProperty.stringValue = "de2c88ca-80fc-448f-bfa9-ab598bf7a9e4";
+            _cloudProjectNameProperty.stringValue = "RS Package Dev Project";
+            _versionProperty.stringValue = "1.3.3.7";
+            _projectSettingsObject.ApplyModifiedProperties();
+        }
+        
+        [UnitySetUp]
+        public IEnumerator Setup()
+        {
+            FixYamatoProjectSettings();
+            
+            var init = UnityServices.InitializeAsync();
+            while (!init.IsCompleted)
+            {
+                yield return null;
+            }
+
+            Task signin = Task.CompletedTask;
+            if (!AuthenticationService.Instance.IsSignedIn)
+            {
+                signin = AuthenticationService.Instance.SignInAnonymouslyAsync();
+            }
+
+            while (!AuthenticationService.Instance.IsSignedIn)
+            {
+                yield return null;
+            }
+
+            ConfigManager.ConfigManagerImpl.FetchConfigs(null, null);
+        }
+
+        [UnityTearDown]
+        public IEnumerator TearDown()
+        {
+            _cloudProjectIdProperty.stringValue = _previousProjectId;
+            _cloudProjectNameProperty.stringValue = _previousProjectName;
+            _versionProperty.stringValue = _previousVersion;
+            _projectSettingsObject.ApplyModifiedProperties();
+
+            yield return null;
+        }
+        
         [UnityTest]
         public IEnumerator SetCustomUserID_SetsCustomUserID()
         {
             ConfigManagerTestUtils.SetProjectIdOnRequestPayload();
             var monoTest = new MonoBehaviourTest<SetCustomUserID_MonobehaviorTest>(false);
 
-            FieldInfo configmanagerImplInfo = typeof(ConfigManager).GetField("ConfigManagerImpl", BindingFlags.Static | BindingFlags.NonPublic);
+            FieldInfo configmanagerImplInfo = typeof(ConfigManager).GetField("_configManagerImpl", BindingFlags.Static | BindingFlags.NonPublic);
             var configmanagerImpl = configmanagerImplInfo.GetValue(null);
 
             monoTest.component.StartTest();
@@ -83,9 +150,10 @@ namespace Unity.RemoteConfig.Tests
         [UnityTest]
         public IEnumerator SaveCache_CreatesFileIfNotThereAndCachesRightContent()
         {
-            if (File.Exists(Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile)))
+            var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+            if (File.Exists(fileName))
             {
-                File.Delete(Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile));
+                File.Delete(fileName);
             }
 
             ConfigManagerTestUtils.SetProjectIdOnRequestPayload();
@@ -93,47 +161,57 @@ namespace Unity.RemoteConfig.Tests
             monoTest.component.StartTest();
             yield return monoTest;
 
-            yield return new WaitForSeconds(2f);
-            Assert.That(File.Exists(Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile)));
-            string text = File.ReadAllText(Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile));
-            Assert.That(text.Contains("testInt"));
-            ConfigManager.ConfigManagerImpl.configs.Remove(ConfigManagerImpl.DefaultConfigKey);
-            ConfigManager.ConfigManagerImpl.LoadFromCache();
-            Assert.That(ConfigManager.appConfig.GetInt("testInt") == 232);
+            yield return waitForFileToBeCreated();
+
+            async Task waitForFileToBeCreated()
+            {
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                string text = File.ReadAllText(fileName);
+                Assert.That(text.Contains("testInt"));
+                ConfigManager.ConfigManagerImpl.configs.Remove(ConfigManagerImpl.DefaultConfigKey);
+                ConfigManager.ConfigManagerImpl.LoadFromCache();
+                Assert.AreEqual(66, ConfigManager.appConfig.GetInt("testInt"));
+            }
+
         }
 
         [UnityTest]
         public IEnumerator SaveCache_CachesMultipleRuntimeConfigs()
         {
-            if (File.Exists(Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile)))
+            var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+            if (File.Exists(fileName))
             {
-                File.Delete(Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile));
+                File.Delete(fileName);
             }
 
             ConfigManagerTestUtils.SetProjectIdOnRequestPayload();
             var monoTest = new MonoBehaviourTest<FetchConfigsComplete_MonobehaviorTest>(false);
             monoTest.component.StartTest();
             yield return monoTest;
-            var managerImpl = ConfigManager.ConfigManagerImpl;
-            var configResponse = managerImpl.ParseResponse(
-                ConfigOrigin.Remote,
-                new Dictionary<string, string>(),
-                ConfigManagerTestUtils.jsonPayloadEconomyConfig);
-            managerImpl.HandleConfigResponse("economy", configResponse);
-            yield return new WaitForSeconds(2f);
-            Assert.That(File.Exists(Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile)));
-            string text = File.ReadAllText(Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile));
-            Assert.That(text.Contains("testInt"));
-            Assert.That(text.Contains("economy"));
-            Assert.That(managerImpl.configs.Count == 2);
-            Assert.That(ConfigManager.GetConfig("economy").GetString("item") == "sword");
-            managerImpl.configs.Remove(ConfigManagerImpl.DefaultConfigKey);
-            managerImpl.configs.Remove("economy");
-            managerImpl.LoadFromCache();
-            Assert.That(ConfigManager.appConfig.GetInt("testInt") == 232);
-            Assert.That(ConfigManager.GetConfig("economy").GetString("item") == "sword");
-            Assert.That(ConfigManager.GetConfig("economy").assignmentId ==
-                        JObject.Parse(ConfigManagerTestUtils.jsonPayloadEconomyConfig)["metadata"]["assignmentId"].ToString());
+
+            yield return waitForFileWithMultipleConfigsToBeCreated();
+
+            async Task waitForFileWithMultipleConfigsToBeCreated()
+            {
+                await Task.Run(() => File.Exists(fileName));
+                var managerImpl = ConfigManager.ConfigManagerImpl;
+                var configResponse = managerImpl.ParseResponse(ConfigOrigin.Remote,new Dictionary<string, string>(),ConfigManagerTestUtils.jsonPayloadEconomyConfig);
+                managerImpl.HandleConfigResponse("economy", configResponse);
+
+                Assert.That(File.Exists(fileName));
+                string text = File.ReadAllText(fileName);
+                Assert.That(text.Contains("testInt"));
+                Assert.That(text.Contains("economy"));
+                Assert.That(managerImpl.configs.Count == 2);
+                Assert.That(ConfigManager.GetConfig("economy").GetString("item") == "sword");
+                managerImpl.configs.Remove(ConfigManagerImpl.DefaultConfigKey);
+                managerImpl.configs.Remove("economy");
+                managerImpl.LoadFromCache();
+                Assert.AreEqual(66, ConfigManager.appConfig.GetInt("testInt"));
+                Assert.That(ConfigManager.GetConfig("economy").GetString("item") == "sword");
+                Assert.That(ConfigManager.GetConfig("economy").assignmentId == JObject.Parse(ConfigManagerTestUtils.jsonPayloadEconomyConfig)["metadata"]["assignmentId"].ToString()); 
+            }
         }
     }
 
@@ -143,27 +221,46 @@ namespace Unity.RemoteConfig.Tests
         public IEnumerator ResponseParsedEventHanlder_ProperlySetsAssignmentId()
         {
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadString);
-
-            yield return null;
-            Assert.That(ConfigManager.appConfig.assignmentId == "a04fb7ec-26e4-4247-b8b4-70dd6967a858");
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual("a04fb7ec-26e4-4247-b8b4-70dd6967a858", ConfigManager.appConfig.assignmentId);
+            }
         }
         [UnityTest]
         public IEnumerator ResponseParsedEventHanlder_ReturnsNullAssignmentIdWhenBadResponse()
         {
             var assignmentIdBeforeRequest = ConfigManager.appConfig.assignmentId;
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadStringNoRCSection);
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual(assignmentIdBeforeRequest, ConfigManager.appConfig.assignmentId);
 
-            yield return null;
-            Assert.That(ConfigManager.appConfig.assignmentId == assignmentIdBeforeRequest);
+            }
         }
 
         [UnityTest]
         public IEnumerator ResponseParsedEventHandler_ProperlySetsEnvironmentId()
         {
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadString);
-
-            yield return null;
-            Assert.That(ConfigManager.appConfig.environmentId == "7d2e0e2d-4bcd-4b6e-8d5d-65d17a708db2");
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual("7d2e0e2d-4bcd-4b6e-8d5d-65d17a708db2", ConfigManager.appConfig.environmentId);
+            }
         }
 
         [UnityTest]
@@ -171,18 +268,30 @@ namespace Unity.RemoteConfig.Tests
         {
             var environmentIdBeforeRequest = ConfigManager.appConfig.environmentId;
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadStringNoRCSection);
-
-            yield return null;
-            Assert.That(ConfigManager.appConfig.environmentId == environmentIdBeforeRequest);
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual(environmentIdBeforeRequest, ConfigManager.appConfig.environmentId);
+            }
         }
 
         [UnityTest]
         public IEnumerator ResponseParsedEventHandler_ProperlySetsConfig()
         {
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadString);
-
-            yield return null;
-            Assert.That(ConfigManager.appConfig.config.ToString() == ((JObject)JObject.Parse(ConfigManagerTestUtils.jsonPayloadString)["configs"]?["settings"])?.ToString());
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual(((JObject)JObject.Parse(ConfigManagerTestUtils.jsonPayloadString)["configs"]?["settings"])?.ToString(), ConfigManager.appConfig.config.ToString());
+            }
         }
 
         [UnityTest]
@@ -190,17 +299,30 @@ namespace Unity.RemoteConfig.Tests
         {
             var configBeforeRequest = ConfigManager.appConfig.config;
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadStringNoRCSection);
-
-            yield return null;
-            Assert.That(ConfigManager.appConfig.config.ToString() == configBeforeRequest.ToString());
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual(configBeforeRequest.ToString(), ConfigManager.appConfig.config.ToString());
+            }
         }
 
         [UnityTest]
         public IEnumerator GetBool_ReturnsRightValue()
         {
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadString);
-            yield return null;
-            Assert.That(ConfigManager.appConfig.GetBool("bool") == true);
+            yield return waitForFileToBeCreated();
+
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.That(ConfigManager.appConfig.GetBool("bool") == true);
+            }        
         }
 
         [UnityTest]
@@ -208,82 +330,151 @@ namespace Unity.RemoteConfig.Tests
         {
             var boolBeforeRequest = ConfigManager.appConfig.GetBool("bool");
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadStringNoRCSection);
-            yield return null;
-            Assert.That(ConfigManager.appConfig.GetBool("bool") == boolBeforeRequest);
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual(boolBeforeRequest, ConfigManager.appConfig.GetBool("bool"));
+            }
         }
 
         [UnityTest]
         public IEnumerator GetFloat_ReturnsRightValue()
         {
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadString);
-            yield return null;
-            Assert.That(ConfigManager.appConfig.GetFloat("helloe") == 0.12999999523162842);
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual(0.12999999523162842, ConfigManager.appConfig.GetFloat("helloe"));
+            }
         }
        [UnityTest]
         public IEnumerator GetFloat_ReturnsRightValueWhenBadResponse()
         {
             var floatBeforeRequest = ConfigManager.appConfig.GetFloat("helloe");
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadStringNoRCSection);
-            yield return null;
-            Assert.That(ConfigManager.appConfig.GetFloat("helloe") == floatBeforeRequest);
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual(floatBeforeRequest, ConfigManager.appConfig.GetFloat("helloe"));
+            }
         }
 
         [UnityTest]
         public IEnumerator GetInt_ReturnsRightValue()
         {
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadString);
-            yield return null;
-            Assert.That(ConfigManager.appConfig.GetInt("someInt") == 12);
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual(12, ConfigManager.appConfig.GetInt("someInt"));
+            }
         }
        [UnityTest]
         public IEnumerator GetInt_ReturnsRightValueWhenBadResponse()
         {
             var intBeforeRequest = ConfigManager.appConfig.GetFloat("someInt");
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadStringNoRCSection);
-            yield return null;
-            Assert.That(ConfigManager.appConfig.GetInt("someInt") == intBeforeRequest);
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual(intBeforeRequest, ConfigManager.appConfig.GetInt("someInt"));
+            }
         }
 
         [UnityTest]
         public IEnumerator GetString_ReturnsRightValue()
         {
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadString);
-            yield return null;
-            Assert.That(ConfigManager.appConfig.GetString("madBro") == "madAF");
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual("madAF", ConfigManager.appConfig.GetString("madBro"));
+            }
         }
 
         [UnityTest]
         public IEnumerator GetString_ReturnsRightValueIfFormattedAsDate()
         {
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadString);
-            yield return null;
-            Assert.That(ConfigManager.appConfig.GetString("stringFormattedAsDate") == "2020-04-03T10:01:00Z");
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual("2020-04-03T10:01:00Z", ConfigManager.appConfig.GetString("stringFormattedAsDate"));
+            }
         }
 
         [UnityTest]
         public IEnumerator GetString_ReturnsRightValueIfFormattedAsJson()
         {
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadString);
-            yield return null;
-            Assert.That(ConfigManager.appConfig.GetString("stringFormattedAsJson") == "{\"a\":2.0,\"b\":4,\"c\":\"someString\"}");
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual("{\"a\":2.0,\"b\":4,\"c\":\"someString\"}", ConfigManager.appConfig.GetString("stringFormattedAsJson"));
+            }
         }
-
-
+        
         [UnityTest]
         public IEnumerator GetString_ReturnsRightValueWhenBadResponse()
         {
             var stringBeforeRequest = ConfigManager.appConfig.GetString("madBro");
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadStringNoRCSection);
-            yield return null;
-            Assert.That(ConfigManager.appConfig.GetString("madBro") == stringBeforeRequest);
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual("madAF", ConfigManager.appConfig.GetString("madBro"));
+            }
         }
 
         [UnityTest]
         public IEnumerator GetLong_ReturnsRightValue()
         {
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadString);
-            yield return null;
-            Assert.That(ConfigManager.appConfig.GetLong("longSomething") == 9223372036854775806);
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual(9223372036854775806, ConfigManager.appConfig.GetLong("longSomething"));
+            }
         }
 
         [UnityTest]
@@ -291,33 +482,61 @@ namespace Unity.RemoteConfig.Tests
         {
             var longBeforeRequest = ConfigManager.appConfig.GetLong("longSomething");
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadStringNoRCSection);
-            yield return null;
-            Assert.That(ConfigManager.appConfig.GetLong("longSomething") == longBeforeRequest);
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual(longBeforeRequest, ConfigManager.appConfig.GetLong("longSomething"));
+            }
         }
 
         [UnityTest]
          public IEnumerator GetJson_ReturnsRightValue()
          {
               ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadString);
-              yield return null;
-              Assert.That(ConfigManager.appConfig.GetJson("jsonSetting") == "{\"a\":1.0,\"b\":2,\"c\":\"someString\"}");
+              yield return waitForFileToBeCreated();
+              
+              async Task waitForFileToBeCreated()
+              {
+                  var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                  await Task.Run(() => File.Exists(fileName));
+                  Assert.That(File.Exists(fileName));
+                  Assert.AreEqual("{\"a\":1.0,\"b\":2,\"c\":\"someString\"}", ConfigManager.appConfig.GetJson("jsonSetting"));
+              }              
          }
 
          [UnityTest]
          public IEnumerator GetJson_ReturnsRightValueWhenBadResponse()
          {
-              var jsonBeforeRequest = ConfigManager.appConfig.GetJson("jsonSetting");
-              ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadStringNoRCSection);
-              yield return null;
-              Assert.That(ConfigManager.appConfig.GetJson("jsonSetting") == jsonBeforeRequest);
+             var jsonBeforeRequest = ConfigManager.appConfig.GetJson("jsonSetting");
+             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadStringNoRCSection);
+
+             yield return waitForFileToBeCreated();
+             
+             async Task waitForFileToBeCreated()
+             {
+                 var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                 await Task.Run(() => File.Exists(fileName));
+                 Assert.That(File.Exists(fileName));
+                 Assert.AreEqual(jsonBeforeRequest, ConfigManager.appConfig.GetJson("jsonSetting") );
+             }
          }
 
         [UnityTest]
         public IEnumerator HasKey_ReturnsRightValue()
         {
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadString);
-            yield return null;
-            Assert.That(ConfigManager.appConfig.HasKey("longSomething"));
+            yield return waitForFileToBeCreated();
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.That(ConfigManager.appConfig.HasKey("longSomething"));
+            }
         }
 
         [UnityTest]
@@ -325,16 +544,30 @@ namespace Unity.RemoteConfig.Tests
         {
             var hasKeylongBeforeRequest = ConfigManager.appConfig.HasKey("longSomething");
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadStringNoRCSection);
-            yield return null;
-            Assert.That(ConfigManager.appConfig.HasKey("longSomething") == hasKeylongBeforeRequest);
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual(hasKeylongBeforeRequest, ConfigManager.appConfig.HasKey("longSomething"));
+            }
         }
 
         [UnityTest]
         public IEnumerator GetKeys_ReturnsRightValue()
         {
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadString);
-            yield return null;
-            Assert.That(ConfigManager.appConfig.GetKeys().Length == ((JObject)JObject.Parse(ConfigManagerTestUtils.jsonPayloadString)["configs"]["settings"]).Properties().Select(prop => prop.Name).ToArray<string>().Length);
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual(((JObject)JObject.Parse(ConfigManagerTestUtils.jsonPayloadString)["configs"]["settings"]).Properties().Select(prop => prop.Name).ToArray<string>().Length, ConfigManager.appConfig.GetKeys().Length);
+            }
         }
 
         [UnityTest]
@@ -342,10 +575,17 @@ namespace Unity.RemoteConfig.Tests
         {
             var keyLengthBeforeRequest = ConfigManager.appConfig.GetKeys().Length;
             ConfigManagerTestUtils.SendPayloadToConfigManager(ConfigManagerTestUtils.jsonPayloadStringNoRCSection);
-            yield return null;
-            Assert.That(ConfigManager.appConfig.GetKeys().Length == keyLengthBeforeRequest);
+            yield return waitForFileToBeCreated();
+            
+            async Task waitForFileToBeCreated()
+            {
+                var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+                await Task.Run(() => File.Exists(fileName));
+                Assert.That(File.Exists(fileName));
+                Assert.AreEqual(keyLengthBeforeRequest, ConfigManager.appConfig.GetKeys().Length );
+            }
         }
-   }
+    }
 
     internal static class ConfigManagerTestUtils
     {
@@ -418,8 +658,12 @@ namespace Unity.RemoteConfig.Tests
                 }
             }";
 
-        public static void SendPayloadToConfigManager(string payload)
+        public async static void SendPayloadToConfigManager(string payload)
         {
+            var fileName = Path.Combine(Application.persistentDataPath, ConfigManagerImpl.DefaultCacheFile);
+
+            await Task.Run(() => File.Exists(fileName));
+            Assert.That(File.Exists(fileName));
 
             ConfigManager.ConfigManagerImpl.HandleConfigResponse(
                 "settings", 
@@ -435,7 +679,7 @@ namespace Unity.RemoteConfig.Tests
 
         public static void SetProjectIdOnRequestPayload()
         {
-            FieldInfo configmanagerImplInfo = typeof(ConfigManager).GetField("ConfigManagerImpl", BindingFlags.Static | BindingFlags.NonPublic);
+            FieldInfo configmanagerImplInfo = typeof(ConfigManager).GetField("_configManagerImpl", BindingFlags.Static | BindingFlags.NonPublic);
             var configmanagerImpl = configmanagerImplInfo.GetValue(null);
 
             FieldInfo rcRequestFieldInfo = typeof(ConfigManagerImpl).GetField("_remoteConfigRequest", BindingFlags.Instance | BindingFlags.NonPublic);